requirements.txt
pyzbar
Appium-Python-Client==0.36
lxml==4.3.1
Pillow==5.4.1
requests==2.21.0
